+++ apps/scoutgame/components/welcome/builder/components/ExtraDetailsForm.tsx
@@ -16,7 +16,7 @@ import {
 } from '@mui/material';
 import type { SessionUser } from '@packages/scoutgame/session/interfaces';
 import { FormErrors } from '@packages/scoutgame-ui/components/common/FormErrors';
-import { EditableUserProfile } from '@packages/scoutgame-ui/components/common/Profile/EditableUserProfile';
+import { EditableUserProfile } from '@packages/scoutgame-ui/components/common/Profile/EditableUserProfile/EditableUserProfile';
 import { useIsMounted } from '@packages/scoutgame-ui/hooks/useIsMounted';
 import { useUser } from '@packages/scoutgame-ui/providers/UserProvider';
 import { concatenateStringValues } from '@root/lib/utils/strings';
