+++ packages/scoutgame-ui/src/components/common/Profile/EditableUserProfile/EditableDisplayname.tsx
@@ -1,0 +1,112 @@
+import CancelIcon from '@mui/icons-material/Cancel';
+import CheckCircleIcon from '@mui/icons-material/CheckCircle';
+import EditIcon from '@mui/icons-material/Edit';
+import { Stack, TextField, Typography } from '@mui/material';
+import { useRef, useState } from 'react';
+import type { Control } from 'react-hook-form';
+import { useController } from 'react-hook-form';
+
+import { ProfileLinks } from '../ProfileLinks';
+
+type EditableDisplayNameProps = {
+  displayName: string;
+  onDisplayNameChange?: (displayName: string) => void;
+  control: Control<any>;
+  isLoading?: boolean;
+  githubLogin?: string;
+  farcasterName?: string | null;
+};
+
+export function EditableDisplayName({
+  displayName,
+  onDisplayNameChange,
+  control,
+  isLoading,
+  githubLogin,
+  farcasterName
+}: EditableDisplayNameProps) {
+  const {
+    field: displayNameField,
+    fieldState: { error }
+  } = useController({
+    name: 'displayName',
+    control
+  });
+  const [isEditingDisplayName, setIsEditingDisplayName] = useState(false);
+  const [isDisplayNameDirty, setIsDisplayNameDirty] = useState(false);
+  const previousDisplayNameRef = useRef<string>(displayName);
+
+  const updateDisplayName = (value: string) => {
+    setIsEditingDisplayName(false);
+    setIsDisplayNameDirty(false);
+    onDisplayNameChange?.(value);
+    previousDisplayNameRef.current = value;
+  };
+
+  const resetDisplayName = () => {
+    setIsEditingDisplayName(false);
+    setIsDisplayNameDirty(false);
+    displayNameField.onChange(previousDisplayNameRef.current);
+  };
+
+  return (
+    <Stack direction='row' alignItems='center' flexWrap='nowrap' justifyContent='space-between' width='100%' gap={1}>
+      <Stack width={isEditingDisplayName ? '80%' : '100%'}>
+        {isEditingDisplayName ? (
+          <TextField
+            required
+            disabled={isLoading}
+            onKeyDown={(e) => {
+              if (e.key === 'Enter') {
+                updateDisplayName(displayNameField.value);
+              } else if (e.key === 'Escape') {
+                resetDisplayName();
+              }
+            }}
+            {...displayNameField}
+            onChange={(event) => {
+              setIsDisplayNameDirty(true);
+              displayNameField.onChange(event);
+            }}
+            fullWidth
+            autoFocus
+            error={!!error?.message}
+            sx={{
+              my: 0.5,
+              '& .MuiInputBase-input': {
+                padding: 0.5,
+                paddingLeft: 1
+              }
+            }}
+          />
+        ) : (
+          <Stack direction='row' alignItems='center' gap={1}>
+            <Typography variant='h6'>{displayNameField.value}</Typography>
+            <EditIcon
+              onClick={() => setIsEditingDisplayName(true)}
+              sx={{ cursor: 'pointer' }}
+              color='primary'
+              fontSize='small'
+            />
+            {farcasterName || githubLogin ? (
+              <ProfileLinks farcasterName={farcasterName} githubLogin={githubLogin} />
+            ) : null}
+          </Stack>
+        )}
+      </Stack>
+      {isEditingDisplayName && (
+        <Stack direction='row' gap={0.5}>
+          {isDisplayNameDirty && !error?.message && displayNameField.value ? (
+            <CheckCircleIcon
+              sx={{ cursor: 'pointer' }}
+              color='success'
+              fontSize='small'
+              onClick={() => updateDisplayName(displayNameField.value)}
+            />
+          ) : null}
+          <CancelIcon sx={{ cursor: 'pointer' }} color='error' fontSize='small' onClick={resetDisplayName} />
+        </Stack>
+      )}
+    </Stack>
+  );
+}
